fix: don't export HeliaRemotePinner class
